+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## [0.0.1] - TODO: Add release date.
-
-* TODO: Describe initial release.
+## 0.0.1
+- Plots datatables
+- Styling, editable, and customization options
+## 1.0.0
+- Major bug fixes
+- Validator now easily integrated
+- Deprecated callable method
+## 1.1.0
+- Massive 16 different validations out-of-the-box, courtesy of FluttableVal
+- FluttableValCheck is a complementing class to help customize validation rules
+- Removed callable method in favour of onTableEdited, onRowAdded, and onRowDeleted
+- Various issues resolved including faulty row data when a row was added